refactor: Add typescript to project
Refactors App, Home and dependent components to use TS.
