+++ package.json
@@ -4,11 +4,20 @@
   "private": true,
   "dependencies": {
     "@turf/helpers": "^6.1.4",
+    "@types/jest": "^24.0.23",
+    "@types/node": "^12.12.11",
+    "@types/react": "^16.9.11",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-router-dom": "^5.1.2",
+    "@types/react-slick": "^0.23.4",
+    "@types/styled-components": "^4.4.0",
+    "eslint-config-react": "^1.1.7",
+    "eslint-config-react-app": "^5.0.2",
     "husky": "^3.0.0",
     "i18next": "^17.0.6",
     "i18next-browser-languagedetector": "^3.0.1",
     "lint-staged": "^9.1.0",
-    "prettier": "1.18.2",
+    "prettier": "^1.19.1",
     "prop-types": "^15.7.2",
     "query-string": "^6.8.1",
     "react": "^16.8.6",
@@ -23,7 +32,8 @@
     "react-slick": "^0.24.0",
     "sanitize.css": "^11.0.0",
     "slick-carousel": "^1.8.1",
-    "supercluster": "^6.0.1"
+    "supercluster": "^6.0.1",
+    "typescript": "^3.7.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -62,10 +72,13 @@
     ]
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "@typescript-eslint/parser": "^2.8.0",
     "commitizen": "^3.1.1",
     "cz-conventional-changelog": "^2.1.0",
     "deepmerge": "^4.0.0",
-    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-plugin-prettier": "^3.1.1",
     "mathjs": "^6.0.3",
     "node-sass": "^4.12.0",
     "pretty-quick": "^1.11.1",
